requirements.txt
alembic==0.9.5
amqp==2.2.1
billiard==3.5.0.3
celery==4.1.0
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
decorator==4.1.2
Flask==0.12.2
Flask-Migrate==2.1.0
flask-oidc==1.1.1
Flask-SQLAlchemy==2.2
httplib2==0.10.3
idna==2.5
itsdangerous==0.24
Jinja2==2.9.6
kombu==4.1.0
Mako==1.0.7
MarkupSafe==1.0
oauth2client==4.1.2
pbr==3.1.1
pyasn1==0.3.2
pyasn1-modules==0.0.11
python-dateutil==2.6.1
python-editor==1.0.3
pytz==2017.2
redis==2.10.6
requests==2.18.3
rsa==3.4.2
six==1.10.0
SQLAlchemy==1.1.13
sqlalchemy-migrate==0.11.0
sqlparse==0.2.3
Tempita==0.5.2
urllib3==1.22
vine==1.1.4
Werkzeug==0.12.2
